extracted `AddonInfo` into separate file / minor `AddonInfo` interface cleanup
